WIP Refactoring MerchangScree
