+++ cardstack/src/screens/index.ts
@@ -1,6 +1,6 @@
 export { default as BuyPrepaidCard } from './BuyPrepaidCard/BuyPrepaidCard';
 export { default as DepotScreen } from './DepotScreen';
-export { default as MerchantScreen } from './MerchantScreen';
+export { default as MerchantScreen } from './MerchantScreen/MerchantScreen';
 export { default as TransactionConfirmation } from './TransactionConfirmation';
 export { default as CopyAddressSheet } from './CopyAddressSheet';
 export { default as PrepaidCardModal } from './PrepaidCardModal';
