+++ src/screens/ExpandedAssetSheet.js
@@ -15,7 +15,7 @@ import {
 import { Centered } from '../components/layout';
 import { useAsset, useDimensions } from '../hooks';
 import { useNavigation } from '../navigation/Navigation';
-import { ExpandedMerchantRoutes } from '@cardstack/screens/MerchantScreen';
+import { ExpandedMerchantRoutes } from '@cardstack/screens/MerchantScreen/MerchantScreen';
 import { position } from '@rainbow-me/styles';
 
 const ScreenTypes = {
